feat: update spin interval to support decimal values and add 30-second testing option

+++ src/main/java/com/remstem/game/model/GameConfig.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class GameConfig {
-    private int spinIntervalMinutes = 15;
+    private double spinIntervalMinutes = 15;
     private Intensity intensity = Intensity.NORMAL;
 }
